Friendship ended with Collada, now glTF 2.0 is my new best friend
This is a huge overhaul of how I export models, it's still a bit messy, but it (mostly) works as expected. Current problems that I'm aware of are some models don't seem to be exporting their full mesh set or node tree, or have multiple meshes attached to single nodes. Either way, the current method I use to attach meshes to nodes is very flawed and should be replaced ASAP.